pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aclarai-claimify"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml",
    "dspy-ai",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
optimization = [
    "dspy-ai>=2.4.0",
    "litellm",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
    "sentence-transformers",
    "faiss-cpu",
]
# Dependencies for the scout agent
scout = [
    "langgraph",
    "langgraph-checkpoint-sqlite",
    "langchain-community",
    "langchain-litellm",
    "langchain-tavily",
    "langchain-google-community",
    "attachments",
    "pydantic",
    "beautifulsoup4",
    "litellm",
    "langsmith",
    "requests",
    "html5lib",
    "httpx",
    "google-api-python-client",
    "google-search-results",
    "json-repair",
    # TUI dependencies for datascout_tui.py
    "textual>=0.48.0",
    "typer>=0.9.0",
]
# Optional dependency for the scout agent's crawler tool, requires playwright is installed on the system
crawler = ["libcrawler"]

[project.scripts]
aclarai-claimify = "aclarai_claimify.cli:main"
aclarai-claimify-scout = "aclarai_claimify.scout.main:main"

[tool.setuptools]
packages = [
    "aclarai_claimify",
    "aclarai_claimify.components",
    "aclarai_claimify.embeddings",
    "aclarai_claimify.optimization",
    "aclarai_claimify.optimization.tests",
    "aclarai_claimify.prompts",
    "aclarai_claimify.scout",
    "aclarai_claimify.settings",
    "aclarai_claimify.utils",
]
[tool.setuptools.package-data]
"aclarai_claimify.settings" = ["*.yaml"]
"aclarai_claimify" = ["../docs/*.md"]
